Split and move down ISemanticFactsService to shared layer. This service is split into two parts, similar to `ISyntaxFacts` and `ISyntaxFactsService`: 1. `ISemanticFacts`, which is purely semantic query API, is now accessible from shared analyzer and code fixes layer. 2. `ISemanticFactsService`, which is semantic query + code transformation API as well as an `ILanguageService`, is only accessible from shared code fixes layer. It stays as being accessible via GetLanguageService API.
